+++ tests/drmcp/unit/test_routes.py
@@ -631,6 +631,7 @@ class TestMetadataRoute:
         mock_tool_config.enable_code_execution_tools = False
         mock_tool_config.enable_optimization_tools = False
         mock_tool_config.enable_vdb_tools = False
+        mock_tool_config.enable_panels_tools = False
         mock_tool_config.enable_workload_tools = False
 
         mock_config.tool_config = mock_tool_config
@@ -740,6 +741,7 @@ class TestMetadataRoute:
         mock_tool_config.enable_code_execution_tools = False
         mock_tool_config.enable_optimization_tools = False
         mock_tool_config.enable_vdb_tools = False
+        mock_tool_config.enable_panels_tools = False
         mock_tool_config.enable_workload_tools = False
 
         mock_config.tool_config = mock_tool_config
@@ -804,6 +806,7 @@ class TestMetadataRoute:
         mock_tool_config.enable_code_execution_tools = False
         mock_tool_config.enable_optimization_tools = False
         mock_tool_config.enable_vdb_tools = False
+        mock_tool_config.enable_panels_tools = False
         mock_tool_config.enable_workload_tools = False
 
         mock_config.tool_config = mock_tool_config